+++ src/Db/Adapter/Oracle.php
@@ -113,7 +113,7 @@ class Oracle extends AbstractAdapter
         $rows = [];
 
         while (($row = $this->fetch()) != false) {
-            $rows[] = $row;
+            $rows[] = new \ArrayObject($row, \ArrayObject::ARRAY_AS_PROPS);
         }
 
         return $rows;
@@ -160,7 +160,7 @@ class Oracle extends AbstractAdapter
             throw new Exception('Error: The database result resource is not currently set.');
         }
 
-        return oci_fetch_array($this->statement, OCI_RETURN_NULLS+OCI_ASSOC);
+        return new \ArrayObject(oci_fetch_array($this->statement, OCI_RETURN_NULLS+OCI_ASSOC), \ArrayObject::ARRAY_AS_PROPS);
     }
 
     /**