+++ src/Db/Sql/Delete.php
@@ -30,26 +30,34 @@ class Delete extends AbstractSql
 
     /**
      * WHERE predicate object
-     * @var \Pop\Db\Sql\Predicate
+     * @var \Pop\Db\Sql\Where
      */
     protected $where = null;
 
     /**
      * Set the WHERE clause
      *
-     * @param  Predicate $where
-     * @return \Pop\Db\Sql\Predicate
+     * @param  $where
+     * @return \Pop\Db\Sql\Delete
      */
     public function where($where = null)
     {
         if (null !== $where) {
-            $this->where = $where;
+            if ($where instanceof Where) {
+                $this->where = $where;
+            } else {
+                if (null === $this->where) {
+                    $this->where = (new Where($this->sql))->add($where);
+                } else {
+                    $this->where->add($where);
+                }
+            }
         }
         if (null === $this->where) {
-            $this->where = new Predicate($this->sql);
+            $this->where = new Where($this->sql);
         }
 
-        return $this->where;
+        return $this;
     }
 
     /**
@@ -78,6 +86,27 @@ class Delete extends AbstractSql
         }
 
         return $sql;
+    }
+
+    /**
+     * Magic method to access $where property
+     *
+     * @param  string $name
+     * @throws Exception
+     * @return mixed
+     */
+    public function __get($name)
+    {
+        switch (strtolower($name)) {
+            case 'where':
+                if (null === $this->where) {
+                    $this->where = new Where($this->sql);
+                }
+                return $this->where;
+                break;
+            default:
+                throw new Exception('Not a valid property for this object.');
+        }
     }
 
 }