+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Pop PHP Framework Changelog
 
 2.0.0a
 ------
-As of May 27, 2014
+As of June 16, 2014
 
 * PHP 5.4+ Only
 * Composer Support
@@ -21,6 +21,7 @@ As of May 27, 2014
         + Removed the top-level Pop\Db\Db class in favor of direct access to the Db\Adapter classes
         + Removed Escaped adapter and support for non-prepared queries
         + Revised the join() method to be more clear in the Sql\Select class
+        + Upgraded and Improved the Db\Sql component
     - File
         + Completely stripped down to only the upload and checkDuplicate static methods
         + Trimmed Dir class, removed references to old File class
@@ -30,7 +31,7 @@ As of May 27, 2014
         + Added support for JSON language files
 * Removed
     - Compress
-    - Db\Adapter\Mysql
+    - Db\Adapter\Mysql (renamed old Mysqli adapter class to Mysql)
     - Filter\Search
     - Loader
     - Project